refactor: update hero
At this time I decided to have a simple button for call to action. I'll get back to the form for automatic installation in the end.

The hero image needs updating. I left the `<HeroIllustration />` comment to go back in the feature and use it instead of a react `Image` component